Подключил аунтификацию и регистрацию. На эндпоинте регистрации стоит depends. Чтобы создать первого юзера нужно снять депендс. Доработка шаблонов, эндпоинтов и небольшая правка схем обновления сообщений. Внес правки в механизм генерации декодирования токена по согласованию с техлидом. Обьединил файлы main.py. Подтянул из ветки девелоп файл docker-compose. Куча мелких правок...
On branch feature/endpoints
Your branch is up to date with 'origin/feature/endpoints'.

Changes to be committed:
	modified:   ../infra/docker-compose.yaml
	modified:   app/admin_endpoints/constants.py
	modified:   app/admin_endpoints/dashboard.py
	modified:   app/admin_endpoints/group.py
	modified:   app/admin_endpoints/message.py
	modified:   app/admin_endpoints/photo.py
	modified:   app/admin_endpoints/user_tg.py
	modified:   app/core/authentication.py
	modified:   app/core/jwt.py
	deleted:    app/main.py
	modified:   app/routers/auth.py
	modified:   app/schemas/message.py
	renamed:    app/admin_endpoints/static/css/bootstrap.min.css -> app/static/css/bootstrap.min.css
	renamed:    app/admin_endpoints/static/css/styles.css -> app/static/css/styles.css
	renamed:    app/admin_endpoints/static/js/bootstrap.bundle.min.js -> app/static/js/bootstrap.bundle.min.js
	new file:   app/templates/403.html
	modified:   app/templates/404.html
	modified:   app/templates/admin_groups.html
	modified:   app/templates/admin_messages.html
	modified:   app/templates/admin_sidebar.html
	modified:   app/templates/admin_users.html
	modified:   app/templates/create_message.html
	modified:   main.py

Untracked files:
	app/static/photos/
