+++ infra/docker-compose.yaml
@@ -1,0 +1,33 @@
+version: '3.8'
+name: providenie
+
+services:
+  app:
+    build: ../src
+    restart: unless-stopped
+    environment:
+      - RUN_IN_DOCKER=1
+    env_file:
+      - ./.env
+    volumes:
+      - ~:/app/bot/.data/
+    ports:
+      - 8000:8000
+    depends_on:
+      - db
+
+  db:
+    image: postgres:17-alpine
+    restart: unless-stopped
+    env_file:
+      - path: ./.env
+        required: true
+    volumes:
+      - db_data:/var/lib/postgresql/data
+      - db_logs:/var/log/postgresql
+    ports:
+      - 5432:5432
+
+volumes:
+  db_data:
+  db_logs:
