+++ src/app/admin_endpoints/constants.py
@@ -5,4 +5,4 @@ PAGE_SIZE_GE = 1
 PAGE_SIZE_LE = 100
 FIRST_PAGE = 1
 EXTRA_PAGE = 1
-STATIC_DIR = "app/admin_endpoints/static/photos"
+STATIC_DIR = "app/static/photos"
