+++ src/app/routers/auth.py
@@ -5,7 +5,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.config import BASE_DIR
 from app.core.auth import get_current_admin, get_current_user
-from app.core.constants import KEY_ACCESS_TOKEN, KEY_USER_ID
+from app.core.constants import KEY_ACCESS_TOKEN
 from app.core.db import get_async_session
 from app.core.jwt import create_access_token, password_verify
 from app.crud.user_tg import crud_user
@@ -58,9 +58,9 @@ async def login(
             {'request': request, 'message': 'Invalid credentials'},
             status_code=401,
         )
-    token = create_access_token({KEY_USER_ID: user.id})
+    token = create_access_token(str(user.id))
     response = RedirectResponse(
-        url='/auth/dashboard/',
+        url='/admin/',
         status_code=303,
     )
     response.set_cookie(
@@ -72,7 +72,7 @@ async def login(
     return response
 
 
-@router.post('/logout/')
+@router.get('/logout/')
 async def logout(response: Response) -> Response:
     """Выход из системы и удаление токена из cookie.
 
@@ -164,7 +164,7 @@ async def register_user(
         )
 
     await crud_user.create(pydantic_scheme_user=user, session=session)
-    return RedirectResponse(url='/auth/login/', status_code=303)
+    return RedirectResponse(url='/login/', status_code=303)
 
 
 @router.get('/dashboard/', response_class=HTMLResponse)
@@ -180,5 +180,14 @@ async def dashboard(request: Request) -> HTMLResponse:
     """
     return templates.TemplateResponse(
         'dashboard.html',
+        {'request': request},
+    )
+
+
+@router.get('/forbidden/', response_class=HTMLResponse)
+async def forbidden(request: Request) -> HTMLResponse:
+    """Endpoint for 403 page."""
+    return templates.TemplateResponse(
+        '403.html',
         {'request': request},
     )
