+++ src/app/templates/403.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="ru">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>403 Доступ запрещен</title>
+    <link rel="stylesheet" href="/static/main.css">
+</head>
+<body>
+    <div class="error-container">
+        <h1>403 - Доступ запрещен</h1>
+        <p>Извините, у вас нет доступа к этой странице.</p>
+        <p>Авторизуйтесь, пожалуйста, чтобы продолжить.</p>
+        <a href="/auth/login">Перейти к авторизации</a>
+        <a href="/admin">Вернуться на главную</a>
+    </div>
+</body>
+</html>
