+++ src/app/templates/404.html
@@ -10,7 +10,7 @@
     <div class="error-container">
         <h1>404 - Страница не найдена</h1>
         <p>Извините, запрашиваемая страница не найдена.</p>
-        <a href="/">Вернуться на главную</a>
+        <a href="/admin">Вернуться на главную</a>
     </div>
 </body>
 </html>
