+++ src/app/templates/admin_groups.html
@@ -9,29 +9,31 @@
             {% include 'admin_sidebar.html' %}
             <main class="col-md-9 ms-sm-auto col-lg-10 px-md-4" style="margin-bottom: 60px;">
                 <h2>Список групп</h2>
-                <table class="table table-striped">
-                    <thead>
-                        <tr>
-                            <th>ID Группы</th>
-                            <th>Название</th>
-                            <th>Действия</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        {% for group in groups %}
-                        <tr>
-                            <td>{{ group.id }}</td>
-                            <td>{{ group.name }}</td>
-                            <td>
-                                <a href="/admin/groups/{{ group.id }}/edit" class="btn btn-warning btn-sm">Изменить</a>
-                                <form action="/admin/groups/{{ group.id }}/delete" method="post" style="display:inline;">
-                                    <button type="submit" class="btn btn-danger btn-sm">Удалить</button>
-                                </form>
-                            </td>
-                        </tr>
-                        {% endfor %}
-                    </tbody>
-                </table>
+                <div class="table-responsive">
+                  <table class="table table-striped">
+                      <thead>
+                          <tr>
+                              <th>ID Группы</th>
+                              <th>Название</th>
+                              <th>Действия</th>
+                          </tr>
+                      </thead>
+                      <tbody>
+                          {% for group in groups %}
+                          <tr>
+                              <td>{{ group.id }}</td>
+                              <td>{{ group.name }}</td>
+                              <td>
+                                  <a href="/admin/groups/{{ group.id }}/edit" class="btn btn-warning btn-sm">Изменить</a>
+                                  <form action="/admin/groups/{{ group.id }}/delete" method="post" style="display:inline;">
+                                      <button type="submit" class="btn btn-danger btn-sm">Удалить</button>
+                                  </form>
+                              </td>
+                          </tr>
+                          {% endfor %}
+                      </tbody>
+                  </table>
+                </div>
 
                 <nav aria-label="Навигация по страницам">
                     <ul class="pagination">
