+++ src/app/templates/admin_messages.html
@@ -41,6 +41,13 @@
 
                 <nav aria-label="Навигация по страницам">
                     <ul class="pagination">
+                        {% if page > 1 %}
+                        <li class="page-item">
+                            <a class="page-link" href="?page={{ page - 1 }}" aria-label="Предыдущая">
+                                <span aria-hidden="true">&laquo;</span> Назад
+                            </a>
+                        </li>
+                        {% endif %}
                         {% for p in range(1, total_pages + 1) %}
                         <li class="page-item {% if p == page %}active{% endif %}">
                             <a class="page-link" href="?page={{ p }}">{{ p }}</a>
@@ -51,14 +58,6 @@
                         <li class="page-item">
                             <a class="page-link" href="?page={{ page + 1 }}" aria-label="Следующая">
                                 Вперед <span aria-hidden="true">&raquo;</span>
-                            </a>
-                        </li>
-                        {% endif %}
-
-                        {% if page > 1 %}
-                        <li class="page-item">
-                            <a class="page-link" href="?page={{ page - 1 }}" aria-label="Предыдущая">
-                                <span aria-hidden="true">&laquo;</span> Назад
                             </a>
                         </li>
                         {% endif %}
