+++ src/app/templates/admin_sidebar.html
@@ -27,6 +27,11 @@
                 <span>Сообщения</span>
             </a>
         </li>
+        <li class="nav-item">
+            <a class="nav-link text-danger" href="/auth/logout">
+                <i class="fas fa-sign-out-alt"></i>
+                <span>Выход</span>
+            </a>
+        </li>
     </ul>
-    
 </nav>
