+++ src/app/templates/admin_users.html
@@ -20,36 +20,44 @@
                     </select>
                     <button type="submit" class="btn btn-primary">Фильтровать</button>
                 </form>
-
-                <table class="table table-striped">
-                    <thead>
-                        <tr>
-                            <th>ID пользователя</th>
-                            <th>Имя</th>
-                            <th>Группы</th>
-                            <th>Действия</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                        {% for user in users %}
-                        <tr>
-                            <td>{{ user.id }}</td>
-                            <td>{{ user.first_name }} {{ user.last_name }}</td>
-                            <td>
-                                {% for group in user.groups %}
-                                    {{ group.name }}{% if not loop.last %}, {% endif %}
-                                {% endfor %}
-                            </td>
-                            <td>
-                                <a href="/admin/users/{{ user.id }}/edit" class="btn btn-secondary btn-sm">Редактировать</a>
-                            </td>
-                        </tr>
-                        {% endfor %}
-                    </tbody>
-                </table>
+                <div class="table-responsive">
+                  <table class="table table-striped">
+                      <thead>
+                          <tr>
+                              <th>ID пользователя</th>
+                              <th>Имя</th>
+                              <th>Группы</th>
+                              <th>Действия</th>
+                          </tr>
+                      </thead>
+                      <tbody>
+                          {% for user in users %}
+                          <tr>
+                              <td>{{ user.id }}</td>
+                              <td>{{ user.first_name }} {{ user.last_name }}</td>
+                              <td>
+                                  {% for group in user.groups %}
+                                      {{ group.name }}{% if not loop.last %}, {% endif %}
+                                  {% endfor %}
+                              </td>
+                              <td>
+                                  <a href="/admin/users/{{ user.id }}/edit" class="btn btn-secondary btn-sm">Редактировать</a>
+                              </td>
+                          </tr>
+                          {% endfor %}
+                      </tbody>
+                  </table>
+                </div>
 
                 <nav aria-label="Навигация по страницам">
                     <ul class="pagination">
+                        {% if page > 1 %}
+                        <li class="page-item">
+                            <a class="page-link" href="?page={{ page - 1 }}" aria-label="Предыдущая">
+                                <span aria-hidden="true">&laquo;</span> Назад
+                            </a>
+                        </li>
+                        {% endif %}
                         {% for p in range(1, total_pages + 1) %}
                         <li class="page-item {% if p == page %}active{% endif %}">
                             <a class="page-link" href="?page={{ p }}">{{ p }}</a>
@@ -60,14 +68,6 @@
                         <li class="page-item">
                             <a class="page-link" href="?page={{ page + 1 }}" aria-label="Следующая">
                                 Вперед <span aria-hidden="true">&raquo;</span>
-                            </a>
-                        </li>
-                        {% endif %}
-
-                        {% if page > 1 %}
-                        <li class="page-item">
-                            <a class="page-link" href="?page={{ page - 1 }}" aria-label="Предыдущая">
-                                <span aria-hidden="true">&laquo;</span> Назад
                             </a>
                         </li>
                         {% endif %}
