+++ src/app/templates/create_message.html
@@ -16,8 +16,8 @@
                         <textarea id="text" name="text" class="form-control" rows="3" required></textarea>
                     </div>
                      <div class="mb-3">
-                        <label for="send_on" class="form-label">Когда отправить (если не задано, то сейчас):</label>
-                        <input type="datetime-local" id="send_on" name="send_on" class="form-control" required>
+                        <label for="send_on" class="form-label">Когда отправить:</label>
+                        <input type="datetime-local" id="send_on" name="send_on" class="form-control">
                     </div>
                     <div class="mb-3">
                         <label for="group_id" class="form-label">Группы:</label>
